requirements.txt
Django==4.2.3
djangorestframework==3.14.0
black==23.7.0
celery==5.3.1
requests==2.31.0
django-environ==0.11.1
django-stubs==4.2.3
pyright==1.1.363
drf-yasg==1.21.7
django-filter==24.2
django-cors-headers==4.3.1
pytest==8.2.1
pytest-django==4.8.0
coverage==7.5.3
